+++ general/components/Resource/PropertiesList.vue
