+++ general/components/Resource/ResourceSection.vue
