+++ general/nuxt.config.js
@@ -18,7 +18,7 @@ process.env.VUE_APP_TAG = process.env.TAG || process.env.TAG_NAME || "latest";
 process.env.VUE_DIST_DIR = `/${process.env.VUE_APP_PRODUCT}/${process.env.VUE_APP_BRANCH}/${process.env.VUE_APP_TAG}`;
 process.env.VUE_ASSETS_DIR = `${process.env.VUE_DIST_DIR}/_nuxt/`;
 
-process.env.VUE_ONTOLOGY_NAME = process.env.ONTPUB_FAMILY || "iof";
+process.env.VUE_ONTOLOGY_NAME = process.env.ONTPUB_FAMILY || "fibo";
 process.env.VUE_BASE_URL =
   process.env.BASE_URL ||
   "https://spec." +
@@ -120,6 +120,7 @@ export default {
       "~/components/Articles",
       "~/components/chunks",
       "~/components/Ontology",
+      "~/components/Resource",
     ],
   },
 
@@ -222,10 +223,10 @@ export default {
       ? process.env.VUE_RESOURCES_BASE_URL.replace(
           "pistoiaalliance",
           "edmcouncil"
-        ) + "*/api"
+        ) + "master/2022Q1/api"
       : process.env.VUE_BASE_URL +
         process.env.VUE_ONTOLOGY_NAME +
-        "/ontology/*/api",
+        "/ontology/master/2022Q1/api",
   ],
 
   styleResources: {
