+++ general/store/servers.js
@@ -1,7 +1,5 @@
 export default {
   state: () => ({
-    ontoviewerDefaultDomain: '',
-
     searchSegment: `find`,
     ontologySegment: `entity`,
     modulesSegment: `module`,
@@ -10,6 +8,7 @@ export default {
     graphSegment: `graph`,
 
     version: null,
+    versionCompare: null,
 
     searchServer: null,
     ontologyServer: null,
@@ -17,45 +16,63 @@ export default {
     statsServer: null,
     missingImportsServer: null,
     graphServer: null,
+
+    searchServerCompare: null,
+    ontologyServerCompare: null,
+    modulesServerCompare: null,
+    statsServerCompare: null,
+    missingImportsServerCompare: null,
+    graphServerCompare: null,
   }),
   mutations: {
     SET_VERSION(state, { version }) {
       state.version = version;
     },
-    SET_SEARCH_SERVER(state, { searchServer }) {
+    SET_SERVERS(
+      state,
+      {
+        searchServer,
+        ontologyServer,
+        modulesServer,
+        statsServer,
+        missingImportsServer,
+        graphServer,
+      }
+    ) {
       state.searchServer = searchServer;
-    },
-    SET_ONTOLOGY_SERVER(state, { ontologyServer }) {
       state.ontologyServer = ontologyServer;
-    },
-    SET_MODULES_SERVER(state, { modulesServer }) {
       state.modulesServer = modulesServer;
-    },
-    SET_STATS_SERVER(state, { statsServer }) {
       state.statsServer = statsServer;
-    },
-    SET_MISSING_IMPORTS_SERVER(state, { missingImportsServer }) {
       state.missingImportsServer = missingImportsServer;
-    },
-    SET_GRAPH_SERVER(state, { graphServer }) {
       state.graphServer = graphServer;
+    },
+    SET_COMPARE_SERVERS(
+      state,
+      {
+        searchServer,
+        ontologyServer,
+        modulesServer,
+        statsServer,
+        missingImportsServer,
+        graphServer,
+      }
+    ) {
+      state.searchServerCompare = searchServer;
+      state.ontologyServerCompare = ontologyServer;
+      state.modulesServerCompare = modulesServer;
+      state.statsServerCompare = statsServer;
+      state.missingImportsServerCompare = missingImportsServer;
+      state.graphServerCompare = graphServer;
     },
   },
   actions: {
     async updateServers({ commit, state, rootState }, { route, to }) {
       let internalRoute = route;
       let ontoviewerDefaultDomain = rootState.configuration.ontoviewerServerUrl;
+      let version = null;
 
       if (to !== undefined) {
         internalRoute = to;
-      }
-
-      if (internalRoute.query?.domain) {
-        ontoviewerDefaultDomain = internalRoute.query.domain;
-      }
-
-      if (internalRoute.query?.modules) {
-        modulesServer = internalRoute.query.modules;
       }
 
       if (internalRoute.query?.version) {
@@ -69,26 +86,74 @@ export default {
         );
         version = internalRoute.query.version;
       } else {
-        ontoviewerDefaultDomain = ontoviewerDefaultDomain.replace("{version}/", "");
-        ontoviewerDefaultDomain = ontoviewerDefaultDomain.replace("{version}", "");
+        ontoviewerDefaultDomain = ontoviewerDefaultDomain.replace(
+          "{version}/",
+          ""
+        );
+        ontoviewerDefaultDomain = ontoviewerDefaultDomain.replace(
+          "{version}",
+          ""
+        );
         version = null;
       }
 
       let searchServer = ontoviewerDefaultDomain + state.searchSegment;
       let ontologyServer = ontoviewerDefaultDomain + state.ontologySegment;
       let statsServer = ontoviewerDefaultDomain + state.statsSegment;
-      let missingImportsServer = ontoviewerDefaultDomain + state.missingImportsSegment;
+      let missingImportsServer =
+        ontoviewerDefaultDomain + state.missingImportsSegment;
       let modulesServer = ontoviewerDefaultDomain + state.modulesSegment;
       let graphServer = ontoviewerDefaultDomain + state.graphSegment;
-      let version = null;
 
       commit("SET_VERSION", { version });
-      commit("SET_SEARCH_SERVER", { searchServer });
-      commit("SET_ONTOLOGY_SERVER", { ontologyServer });
-      commit("SET_MODULES_SERVER", { modulesServer });
-      commit("SET_STATS_SERVER", { statsServer });
-      commit("SET_MISSING_IMPORTS_SERVER", { missingImportsServer });
-      commit("SET_GRAPH_SERVER", { graphServer });
+      commit("SET_SERVERS", {
+        searchServer,
+        ontologyServer,
+        modulesServer,
+        statsServer,
+        missingImportsServer,
+        graphServer,
+      });
+    },
+    async updateCompareServers({ commit, state, rootState }, { compareVersion }) {
+      let ontoviewerDefaultDomain = rootState.configuration.ontoviewerServerUrl;
+
+      if (compareVersion) {
+        ontoviewerDefaultDomain = ontoviewerDefaultDomain.replace(
+          "{version}/",
+          `${compareVersion}/`
+        );
+        ontoviewerDefaultDomain = ontoviewerDefaultDomain.replace(
+          "{version}",
+          `${compareVersion}`
+        );
+      } else {
+        ontoviewerDefaultDomain = ontoviewerDefaultDomain.replace(
+          "{version}/",
+          ""
+        );
+        ontoviewerDefaultDomain = ontoviewerDefaultDomain.replace(
+          "{version}",
+          ""
+        );
+      }
+
+      let searchServer = ontoviewerDefaultDomain + state.searchSegment;
+      let ontologyServer = ontoviewerDefaultDomain + state.ontologySegment;
+      let statsServer = ontoviewerDefaultDomain + state.statsSegment;
+      let missingImportsServer =
+        ontoviewerDefaultDomain + state.missingImportsSegment;
+      let modulesServer = ontoviewerDefaultDomain + state.modulesSegment;
+      let graphServer = ontoviewerDefaultDomain + state.graphSegment;
+
+      commit("SET_COMPARE_SERVERS", {
+        searchServer,
+        ontologyServer,
+        modulesServer,
+        statsServer,
+        missingImportsServer,
+        graphServer,
+      });
     },
   },
 };
